Loggers weren't logging if they overrode the base class log(int state) - because it had been changed to log(long state). Put a final int version in the base class to prevent overriding.

+++ src/dr/evomodel/speciation/AlloppLeggedTree.java
@@ -480,7 +480,7 @@ public class AlloppLeggedTree implements MutableTree, TreeLogger.LogUpon  {
 
 	// TreeLogger.LogUpon
 	
-	public boolean logNow(int state) {
+	public boolean logNow(long state) {
        //		grjtodo 
 
 		return false;
